refactor: 抽离 field 公共文件
