+++ packages/s2-core/src/common/constant/index.ts
@@ -1,3 +1,4 @@
+export * from './field';
 export * from './events';
 export * from './basic';
 export * from './classnames';
